+++ code/go/0chain.net/smartcontract/storagesc/readpool_test.go
@@ -54,7 +54,7 @@ func requireErrMsg(t *testing.T, err error, msg string) {
 
 func Test_lockRequest_decode(t *testing.T) {
 	var lre, lrd lockRequest
-	lre.Duration = int64(time.Second * 60)
+	lre.Duration = time.Second * 60
 	lre.AllocationID = "alloc_hex"
 	lre.BlobberID = "blobber_hex"
 	require.NoError(t, lrd.decode(mustEncode(t, &lre)))
@@ -217,8 +217,8 @@ func TestStorageSmartContract_readPoolLock(t *testing.T) {
 
 	testSetReadPoolConfig(t, &readPoolConfig{
 		MinLock:       10,
-		MinLockPeriod: int64(10 * time.Second),
-		MaxLockPeriod: int64(100 * time.Second),
+		MinLockPeriod: 10 * time.Second,
+		MaxLockPeriod: 100 * time.Second,
 	}, balances, ssc.ID)
 
 	var fp fundedPools = []string{client.id}
@@ -237,7 +237,7 @@ func TestStorageSmartContract_readPoolLock(t *testing.T) {
 	requireErrMsg(t, err, errMsg2)
 	// 3. min lock
 	tx.Value = 5
-	lr.Duration = int64(5 * time.Second)
+	lr.Duration = 5 * time.Second
 	lr.AllocationID = allocID
 	_, err = ssc.readPoolLock(&tx, mustEncode(t, &lr), balances)
 	requireErrMsg(t, err, errMsg4)
@@ -247,11 +247,11 @@ func TestStorageSmartContract_readPoolLock(t *testing.T) {
 	_, err = ssc.readPoolLock(&tx, mustEncode(t, &lr), balances)
 	requireErrMsg(t, err, errMsg5)
 	// 6. max lock period
-	lr.Duration = int64(150 * time.Second)
+	lr.Duration = 150 * time.Second
 	_, err = ssc.readPoolLock(&tx, mustEncode(t, &lr), balances)
 	requireErrMsg(t, err, errMsg6)
 	// 7. no such allocation
-	lr.Duration = int64(15 * time.Second)
+	lr.Duration = 15 * time.Second
 	_, err = ssc.readPoolLock(&tx, mustEncode(t, &lr), balances)
 	require.Error(t, err)
