+++ code/go/0chain.net/smartcontract/storagesc/writepool_test.go
@@ -157,8 +157,8 @@ func TestStorageSmartContract_writePoolLock(t *testing.T) {
 
 	testSetWritePoolConfig(t, &writePoolConfig{
 		MinLock:       10,
-		MinLockPeriod: int64(20 * time.Second),
-		MaxLockPeriod: int64(2 * time.Hour),
+		MinLockPeriod: 20 * time.Second,
+		MaxLockPeriod: 2 * time.Hour,
 	}, balances, ssc.ID)
 
 	var fp fundedPools = []string{client.id}
@@ -173,7 +173,7 @@ func TestStorageSmartContract_writePoolLock(t *testing.T) {
 			&BlobberAllocation{MinLockDemand: 10, Spent: 0},
 		},
 		Expiration:              10,
-		ChallengeCompletionTime: int64(200 * time.Second),
+		ChallengeCompletionTime: 200 * time.Second,
 		Owner:                   client.id,
 	}
 
@@ -193,7 +193,7 @@ func TestStorageSmartContract_writePoolLock(t *testing.T) {
 	requireErrMsg(t, err, errMsg2)
 	// 3. min lock
 	tx.Value = 5
-	lr.Duration = int64(5 * time.Second)
+	lr.Duration = 5 * time.Second
 	lr.AllocationID = allocID
 	_, err = ssc.writePoolLock(&tx, mustEncode(t, &lr), balances)
 	requireErrMsg(t, err, errMsg4)
@@ -203,11 +203,11 @@ func TestStorageSmartContract_writePoolLock(t *testing.T) {
 	_, err = ssc.writePoolLock(&tx, mustEncode(t, &lr), balances)
 	requireErrMsg(t, err, errMsg5)
 	// 6. max lock period
-	lr.Duration = int64(3 * time.Hour)
+	lr.Duration = 3 * time.Hour
 	_, err = ssc.writePoolLock(&tx, mustEncode(t, &lr), balances)
 	requireErrMsg(t, err, errMsg6)
 	// 7. no such allocation
-	lr.Duration = int64(15 * time.Second)
+	lr.Duration = 15 * time.Second
 	_, err = ssc.writePoolLock(&tx, mustEncode(t, &lr), balances)
 	require.Error(t, err)
